requirements.txt
fastapi~=0.100.0
SQLAlchemy~=2.0.19
starlette~=0.27.0
pydantic~=2.1.1
uvicorn~=0.23.1
passlib
bcrypt
python-jose
python-multipart
python-dotenv